1.4.3
Code cleanup, bug fixes, weapon balance and reworks, QOL additions

+++ Projectiles/Swingy.cs
@@ -39,7 +39,7 @@ namespace Infernus.Projectiles
 
         public override Color? GetAlpha(Color lightColor)
         {
-            return new Color(255, 255, 255, 0) * (1f - (float)Projectile.alpha / 255f);
+            return new Color(255, 255, 255, 0) * (1f - Projectile.alpha / 255f);
         }
 
         public override void AI()
